Move configuration loader implementation to separate file
